Extend serde-extensions test with additional extension and backend
