+++ .github/workflows/ci.yml
@@ -45,7 +45,7 @@ jobs:
         if: matrix.target == 'x86_64-unknown-linux-gnu'
 
       - name: Run tests
-        run: cargo test --verbose --features ext,virt
+        run: cargo test --verbose --features serde-extensions,virt
         if: matrix.target == 'x86_64-unknown-linux-gnu'
 
       - name: Check formatting
