+++ Cargo.toml
@@ -121,5 +121,5 @@ test-attestation-cert-ids = []
 # interchange = { git = "https://github.com/trussed-dev/interchange", branch = "main" }
 
 [package.metadata.docs.rs]
-features = ["serialized-extensions", "virt"]
+features = ["serde-extensions", "virt"]
 rustdoc-args = ["--cfg", "docsrs"]
