+++ bacon.toml
@@ -37,6 +37,11 @@ command = ["cargo", "test", "--color", "always"]
 need_stdout = true
 watch = ["src", "tests"]
 
+[jobs.test-ext]
+command = ["cargo", "test", "--color", "always", "--features", "virt,serde-extensions"]
+need_stdout = true
+watch = ["src", "tests"]
+
 [jobs.doc]
 command = ["cargo", "doc", "--color", "always"]
 need_stdout = true
